# backend/openapi/swagger.yaml
basePath: /
definitions:
...
  domain.AnydocUploadResp:
    properties:
      code:
        type: integer
      data:
        type: string
      err:
        type: string
    type: object
...
  domain.AppType:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    - 9
    - 10
    - 11
    - 12
    format: int32
    type: integer
    x-enum-varnames:
    - AppTypeWeb
    - AppTypeWidget
    - AppTypeDingTalkBot
    - AppTypeFeishuBot
    - AppTypeWechatBot
    - AppTypeWechatServiceBot
    - AppTypeDisCordBot
    - AppTypeWechatOfficialAccount
    - AppTypeOpenAIAPI
    - AppTypeWecomAIBot
    - AppTypeLarkBot
    - AppTypeMcpServer
...
  domain.ChatRequest:
    properties:
      app_type:
        allOf:
        - $ref: '#/definitions/domain.AppType'
        enum:
        - 1
        - 2
      captcha_token:
        type: string
      conversation_id:
        type: string
      image_paths:
        items:
          type: string
        maxItems: 3
        type: array
      message:
        type: string
      nonce:
        type: string
    required:
    - app_type
    type: object
  domain.ChatSearchReq:
    properties:
      captcha_token:
        type: string
      message:
        type: string
    required:
    - message
    type: object
  domain.ChatSearchResp:
    properties:
      node_result:
        items:
          $ref: '#/definitions/domain.NodeContentChunkSSE'
        type: array
    type: object
...
  domain.FeedbackRequest:
    properties:
      conversation_id:
        type: string
      feedback_content:
        description: 限制内容长度
        maxLength: 200
        type: string
      message_id:
        type: string
      score:
        allOf:
        - $ref: '#/definitions/domain.ScoreType'
        description: -1 踩 ,0 1 赞成
      type:
        description: 内容不准确，没有帮助，.......
        type: string
    required:
    - message_id
    type: object
...
  domain.MessageContent:
    type: object
...
  domain.NodeContentChunkSSE:
    properties:
      emoji:
        type: string
      name:
        type: string
      node_id:
        type: string
      node_path_names:
        items:
          type: string
        type: array
      summary:
        type: string
    type: object
...
  domain.ObjectUploadResp:
    properties:
      filename:
        type: string
      key:
        type: string
    type: object
...
  domain.OpenAIChoice:
    properties:
      delta:
        allOf:
        - $ref: '#/definitions/domain.OpenAIMessage'
        description: for streaming
      finish_reason:
        type: string
      index:
        type: integer
      message:
        $ref: '#/definitions/domain.OpenAIMessage'
    type: object
  domain.OpenAICompletionsRequest:
    properties:
      frequency_penalty:
        type: number
      max_tokens:
        type: integer
      messages:
        items:
          $ref: '#/definitions/domain.OpenAIMessage'
        type: array
      model:
        type: string
      presence_penalty:
        type: number
      response_format:
        $ref: '#/definitions/domain.OpenAIResponseFormat'
      stop:
        items:
          type: string
        type: array
      stream:
        type: boolean
      stream_options:
        $ref: '#/definitions/domain.OpenAIStreamOptions'
      temperature:
        type: number
      tool_choice:
        $ref: '#/definitions/domain.OpenAIToolChoice'
      tools:
        items:
          $ref: '#/definitions/domain.OpenAITool'
        type: array
      top_p:
        type: number
      user:
        type: string
    required:
    - messages
    - model
    type: object
  domain.OpenAICompletionsResponse:
    properties:
      choices:
        items:
          $ref: '#/definitions/domain.OpenAIChoice'
        type: array
      created:
        type: integer
      id:
        type: string
      model:
        type: string
      object:
        type: string
      usage:
        $ref: '#/definitions/domain.OpenAIUsage'
    type: object
  domain.OpenAIError:
    properties:
      code:
        type: string
      message:
        type: string
      param:
        type: string
      type:
        type: string
    type: object
  domain.OpenAIErrorResponse:
    properties:
      error:
        $ref: '#/definitions/domain.OpenAIError'
    type: object
  domain.OpenAIFunction:
    properties:
      description:
        type: string
      name:
        type: string
      parameters:
        additionalProperties: true
        type: object
    required:
    - name
    type: object
  domain.OpenAIFunctionCall:
    properties:
      arguments:
        type: string
      name:
        type: string
    required:
    - arguments
    - name
    type: object
  domain.OpenAIFunctionChoice:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  domain.OpenAIMessage:
    properties:
      content:
        $ref: '#/definitions/domain.MessageContent'
      name:
        type: string
      role:
        type: string
      tool_call_id:
        type: string
      tool_calls:
        items:
          $ref: '#/definitions/domain.OpenAIToolCall'
        type: array
    required:
    - role
    type: object
  domain.OpenAIResponseFormat:
    properties:
      type:
        type: string
    required:
    - type
    type: object
  domain.OpenAIStreamOptions:
    properties:
      include_usage:
        type: boolean
    type: object
  domain.OpenAITool:
    properties:
      function:
        $ref: '#/definitions/domain.OpenAIFunction'
      type:
        type: string
    required:
    - type
    type: object
  domain.OpenAIToolCall:
    properties:
      function:
        $ref: '#/definitions/domain.OpenAIFunctionCall'
      id:
        type: string
      type:
        type: string
    required:
    - function
    - id
    - type
    type: object
  domain.OpenAIToolChoice:
    properties:
      function:
        $ref: '#/definitions/domain.OpenAIFunctionChoice'
      type:
        type: string
    type: object
  domain.OpenAIUsage:
    properties:
      completion_tokens:
        type: integer
      prompt_tokens:
        type: integer
      total_tokens:
        type: integer
    type: object
...
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  domain.ScoreType:
    enum:
    - 1
    - -1
    type: integer
    x-enum-varnames:
    - Like
    - DisLike
...
  domain.UploadByUrlReq:
    properties:
      kb_id:
        type: string
      url:
        type: string
    required:
    - url
    type: object
...
  v1.FileUploadResp:
    properties:
      key:
        type: string
    type: object
...
  v1.ShareFileUploadUrlReq:
    properties:
      captcha_token:
        type: string
      url:
        type: string
    required:
    - captcha_token
    - url
    type: object
  v1.ShareFileUploadUrlResp:
    properties:
      key:
        type: string
    type: object
...
info:
  contact: {}
  description: niuniu-wiki API documentation
  title: niuniu-wiki API
  version: "1.0"
paths:
...
  /api/v1/file/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload File
      parameters:
      - description: File
        in: formData
        name: file
        required: true
        type: file
      - description: Knowledge Base ID
        in: formData
        name: kb_id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ObjectUploadResp'
      summary: Upload File
      tags:
      - file
  /api/v1/file/upload/anydoc:
    post:
      consumes:
      - multipart/form-data
      description: Upload Anydoc File
      parameters:
      - description: File
        in: formData
        name: file
        required: true
        type: file
      - description: File Path
        in: formData
        name: path
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AnydocUploadResp'
      summary: Upload Anydoc File
      tags:
      - file
  /api/v1/file/upload/url:
    post:
      consumes:
      - application/json
      description: Upload File By Url
      parameters:
      - description: Request Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UploadByUrlReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ObjectUploadResp'
              type: object
      summary: Upload File By Url
      tags:
      - file
...
  /share/v1/chat/completions:
    post:
      consumes:
      - application/json
      description: OpenAI API compatible chat completions endpoint
      parameters:
      - description: Knowledge Base ID
        in: header
        name: X-KB-ID
        required: true
        type: string
      - description: OpenAI API request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.OpenAICompletionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.OpenAICompletionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.OpenAIErrorResponse'
      summary: ChatCompletions
      tags:
      - share_chat
  /share/v1/chat/feedback:
    post:
      consumes:
      - application/json
      description: Process user feedback for chat conversations
      parameters:
      - description: feedback request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.FeedbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Handle chat feedback
      tags:
      - share_chat
  /share/v1/chat/message:
    post:
      consumes:
      - application/json
      description: ChatMessage
      parameters:
      - description: app type
        in: query
        name: app_type
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: ChatMessage
      tags:
      - share_chat
  /share/v1/chat/search:
    post:
      consumes:
      - application/json
      description: ChatSearch
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ChatSearchReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ChatSearchResp'
              type: object
      summary: ChatSearch
      tags:
      - share_chat_search
  /share/v1/chat/widget:
    post:
      consumes:
      - application/json
      description: ChatWidget
      parameters:
      - description: app type
        in: query
        name: app_type
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: ChatWidget
      tags:
      - Widget
  /share/v1/chat/widget/search:
    post:
      consumes:
      - application/json
      description: WidgetSearch
      parameters:
      - description: Comment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ChatSearchReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ChatSearchResp'
              type: object
      summary: WidgetSearch
      tags:
      - Widget
...
  /share/v1/common/file/upload:
    post:
      consumes:
      - multipart/form-data
      description: 前台用户上传文件,目前只支持图片文件上传
      operationId: share-FileUpload
      parameters:
      - description: kb id
        in: header
        name: X-KB-ID
        required: true
        type: string
      - description: File
        in: formData
        name: file
        required: true
        type: file
      - description: captcha_token
        in: formData
        name: captcha_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.FileUploadResp'
              type: object
      summary: 文件上传
      tags:
      - ShareFile
  /share/v1/common/file/upload/url:
    post:
      consumes:
      - application/json
      description: 前台用户上传文件,目前只支持图片文件上传
      operationId: share-FileUploadByUrl
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.ShareFileUploadUrlReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.ShareFileUploadUrlResp'
              type: object
      summary: 文件上传
      tags:
      - ShareFile
...
securityDefinitions:
  bearerAuth:
    description: Type "Bearer" + a space + your token to authorize
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"